Split Connection rel. controls to a separate component
