+++ assets/source/setup-guide/app/steps/SetupAccount.js
@@ -1,30 +1,18 @@
 /**
  * External dependencies
  */
-import { sprintf, __ } from '@wordpress/i18n';
-import { decodeEntities } from '@wordpress/html-entities';
-import { createInterpolateElement, useState } from '@wordpress/element';
-import apiFetch from '@wordpress/api-fetch';
-import { Spinner } from '@woocommerce/components';
-import { getNewPath } from '@woocommerce/navigation';
-import {
-	Button,
-	Card,
-	CardBody,
-	CardFooter,
-	Flex,
-	FlexItem,
-	FlexBlock,
-	Modal,
-	__experimentalText as Text, // eslint-disable-line @wordpress/no-unsafe-wp-apis --- _experimentalText unlikely to change/disappear and also used by WC Core
-} from '@wordpress/components';
+
+import { createInterpolateElement } from '@wordpress/element';
+import { __ } from '@wordpress/i18n';
+import { Button, Card, CardFooter } from '@wordpress/components';
 
 /**
  * Internal dependencies
  */
 import StepHeader from '../components/StepHeader';
 import StepOverview from '../components/StepOverview';
-import { useSettingsSelect, useCreateNotice } from '../helpers/effects';
+import AccountConnection from '../components/Account/Connection';
+import { useSettingsSelect } from '../helpers/effects';
 
 const SetupAccount = ( {
 	goToNextStep,
@@ -32,86 +20,7 @@ const SetupAccount = ( {
 	isBusinessConnected,
 	setIsBusinessConnected,
 } ) => {
-	const [ isConfirmationModalOpen, setIsConfirmationModalOpen ] = useState(
-		false
-	);
 	const appSettings = useSettingsSelect();
-
-	const createNotice = useCreateNotice();
-
-	const openConfirmationModal = () => {
-		setIsConfirmationModalOpen( true );
-	};
-
-	const closeConfirmationModal = () => {
-		setIsConfirmationModalOpen( false );
-	};
-
-	const renderConfirmationModal = () => {
-		return (
-			<Modal
-				title={
-					<>{ __( 'Are you sure?', 'pinterest-for-woocommerce' ) }</>
-				}
-				onRequestClose={ closeConfirmationModal }
-				className="woocommerce-setup-guide__step-modal"
-			>
-				<div className="woocommerce-setup-guide__step-modal__wrapper">
-					<p>
-						{ __(
-							'Are you sure you want to disconnect this account?',
-							'pinterest-for-woocommerce'
-						) }
-					</p>
-					<div className="woocommerce-setup-guide__step-modal__buttons">
-						<Button
-							isDestructive
-							isSecondary
-							onClick={ handleDisconnectAccount }
-						>
-							{ __(
-								"Yes, I'm sure",
-								'pinterest-for-woocommerce'
-							) }
-						</Button>
-						<Button isTertiary onClick={ closeConfirmationModal }>
-							{ __( 'Cancel', 'pinterest-for-woocommerce' ) }
-						</Button>
-					</div>
-				</div>
-			</Modal>
-		);
-	};
-
-	const handleDisconnectAccount = async () => {
-		closeConfirmationModal();
-
-		const result = await apiFetch( {
-			path:
-				wcSettings.pinterest_for_woocommerce.apiRoute +
-				'/auth_disconnect',
-			method: 'POST',
-		} );
-
-		if ( ! result.disconnected ) {
-			createNotice(
-				'error',
-				__(
-					'There was a problem while trying to disconnect.',
-					'pinterest-for-woocommerce'
-				)
-			);
-		} else {
-			setIsBusinessConnected( false );
-
-			// Force reload WC admin page to initiate the relevant dependencies of the Dashboard page.
-			const path = getNewPath( {}, '/pinterest/landing', {} );
-
-			window.location = new URL(
-				decodeEntities( wcSettings.adminUrl + path )
-			);
-		}
-	};
 
 	return (
 		<div className="woocommerce-setup-guide__setup-account">
@@ -163,72 +72,11 @@ const SetupAccount = ( {
 				</div>
 				<div className="woocommerce-setup-guide__step-column">
 					<Card>
-						<CardBody size="large">
-							{ isBusinessConnected === true ? ( // eslint-disable-line no-nested-ternary --- Code is reasonable readable
-								<Flex>
-									<FlexBlock className="is-connected">
-										<Text variant="subtitle">
-											{ __(
-												'Pinterest Account',
-												'pinterest-for-woocommerce'
-											) }
-										</Text>
-										{ appSettings?.account_data?.id && (
-											<Text variant="body">
-												{ sprintf(
-													'%1$s: %2$s - %3$s',
-													__(
-														'Account',
-														'pinterest-for-woocommerce'
-													),
-													appSettings.account_data
-														.username,
-													appSettings.account_data.id
-												) }
-											</Text>
-										) }
-										<Button
-											isLink
-											isDestructive
-											onClick={ openConfirmationModal }
-										>
-											{ __(
-												'Disconnect Pinterest Account',
-												'pinterest-for-woocommerce'
-											) }
-										</Button>
-									</FlexBlock>
-								</Flex>
-							) : isBusinessConnected === false ? (
-								<Flex>
-									<FlexBlock>
-										<Text variant="subtitle">
-											{ __(
-												'Connect your Pinterest Account',
-												'pinterest-for-woocommerce'
-											) }
-										</Text>
-									</FlexBlock>
-									<FlexItem>
-										<Button
-											isSecondary
-											href={ decodeEntities(
-												wcSettings
-													.pinterest_for_woocommerce
-													.serviceLoginUrl
-											) }
-										>
-											{ __(
-												'Connect',
-												'pinterest-for-woocommerce'
-											) }
-										</Button>
-									</FlexItem>
-								</Flex>
-							) : (
-								<Spinner />
-							) }
-						</CardBody>
+						<AccountConnection
+							isBusinessConnected={ isBusinessConnected }
+							setIsBusinessConnected={ setIsBusinessConnected }
+							accountData={ appSettings.account_data }
+						/>
 
 						{ isBusinessConnected === false && (
 							<CardFooter>
@@ -247,8 +95,6 @@ const SetupAccount = ( {
 								</Button>
 							</CardFooter>
 						) }
-
-						{ isConfirmationModalOpen && renderConfirmationModal() }
 					</Card>
 
 					{ view === 'wizard' && isBusinessConnected === true && (
